requirements.txt
asgiref==3.2.10
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
boto3==1.15.11
botocore==1.18.11
cffi==1.14.3
channels==2.4.0
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
Django==3.1.2
django-annoying==0.10.6
django-crispy-forms==1.9.2
django-extra-views==0.13.0
django-gravatar2==1.4.4
django-model-utils==4.0.0
django-picklefield==3.0.1
django-ses==1.0.3
django-tables2==2.3.1
djangorestframework==3.12.1
future==0.18.2
gameservice==1.4.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jmespath==0.10.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
treys==0.1.3
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.10
zope.interface==5.1.2
